Consolidating some sprite names
Whoops! I missed some differences in the sprite names that kept the program from running. (There was still some references to the player as "ship"). I've decided to favor the newer "player.Sprite" that I used in my own program (where the texture is loaded as a part of the player class instead of in the Game class, making it easier if we want to change later) but if that interferes with other people's code, it's just as subject to change.